+++ src/main/kotlin/com/tsukanov/news/KotlinSpringExampleApplication.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinspringexample
+package com.tsukanov.news
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
