+++ src/main/kotlin/com/tsukanov/news/entity/Keyword.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinspringexample.entity
+package com.tsukanov.news.entity
 
 import javax.persistence.*
