+++ src/main/kotlin/com/tsukanov/news/repository/ArticleRepository.kt
@@ -1,6 +1,6 @@
-package com.example.kotlinspringexample.repository
+package com.tsukanov.news.repository
 
-import com.example.kotlinspringexample.entity.Article
+import com.tsukanov.news.entity.Article
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 import org.springframework.data.repository.PagingAndSortingRepository
 import org.springframework.stereotype.Repository
