+++ src/test/kotlin/com/tsukanov/news/KotlinSpringExampleApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinspringexample
+package com.tsukanov.news
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
